refactor(client): split monolithic tauri.ts into domain modules
Split 4697-line lib/tauri.ts into 15 domain files under lib/tauri/:
common (shared state, HTTP helpers, upload auth, file validation),
auth, channels, messages, guilds, voice, social, admin, settings,
pages, e2ee, websocket, search, dms. index.ts re-exports all so
existing `@/lib/tauri` imports continue to work.
